Various fixes for embedding
