+++ panel/pane/base.py
@@ -8,10 +8,10 @@ from functools import partial
 
 import param
 
-from ..io import state
+from ..io import push, state
 from ..layout import Panel, Row
 from ..viewable import Viewable, Reactive, Layoutable
-from ..util import param_reprs, push
+from ..util import param_reprs
 
 
 def Pane(obj, **kwargs):
